Break cyclic imports for CodeQL
Move shared GPU helpers and kernel decision settings below their callers so orchestrator runtime modules no longer import through CLI or SharedState during import analysis.
