+++ src/hyperloom/inference_optimizer/gpu_types.py
@@ -1,0 +1,82 @@
+# Copyright Advanced Micro Devices, Inc. All rights reserved.
+
+"""AMD GPU-type helpers shared by the CLI and orchestrator runtime."""
+
+from __future__ import annotations
+
+import os
+
+
+_AMD_GPU_TYPES = frozenset({"mi300x", "mi308x", "mi325x", "mi355x"})
+
+_GFX_TO_RUNNER: dict[str, str] = {
+    # Mirror Magpie/modes/benchmark/image_selector.py:138-140 so launchers and
+    # runtime materializers agree on the selected benchmark script.
+    "gfx942": "mi300x",
+    "gfx950": "mi355x",
+}
+
+
+def _gpu_runner_type(gpu_type: str) -> str:
+    """Return the Magpie runner label for a resolved real GPU type."""
+    normalized = str(gpu_type or "").strip().lower()
+    if normalized in ("mi325x", "mi308x"):
+        return "mi300x"
+    return normalized
+
+
+def _resolve_gpu_type(
+    user_specified: str,
+    probed: str,
+) -> tuple[str, list[str]]:
+    """Resolve effective gpu_type from a user hint and a hardware probe."""
+    warnings: list[str] = []
+    if probed and user_specified and probed != user_specified:
+        warnings.append(
+            f"WARN: --gpu-type={user_specified!r} disagrees with probed "
+            f"{probed!r}; using probed {probed!r}. The probe wins because "
+            f"Magpie runner_type + KB recipe rows must match the actual "
+            f"hardware to keep baseline numbers comparable across sessions."
+        )
+        return probed, warnings
+    return (probed or user_specified), warnings
+
+
+def _autodetect_gpu_type() -> str | None:
+    """Return mi300x|mi308x|mi325x|mi355x or None if undetectable."""
+    import subprocess
+
+    try:
+        out = subprocess.run(
+            ["rocm-smi", "--showproductname"],
+            capture_output=True,
+            text=True,
+            timeout=5,
+        ).stdout.upper()
+        for tag in ("MI355X", "MI325X", "MI308X", "MI300X"):
+            if tag in out:
+                return tag.lower()
+    except (FileNotFoundError, subprocess.TimeoutExpired, PermissionError, OSError):
+        # rocm-smi missing / slow / not permitted; fall through to the torch
+        # gcnArchName probe below (autodetect is best-effort).
+        pass
+    try:
+        import torch
+
+        arch = torch.cuda.get_device_properties(0).gcnArchName
+        gfx = arch.split(":", 1)[0].lower()
+        return _GFX_TO_RUNNER.get(gfx)
+    except Exception:  # noqa: BLE001
+        return None
+
+
+def _resolve_amd_gpu_type(explicit: str | None = None) -> str | None:
+    """Resolve the current AMD GPU type, or None when not on AMD/unknown."""
+    explicit_norm = str(explicit or "").strip().lower()
+    if explicit_norm:
+        return explicit_norm if explicit_norm in _AMD_GPU_TYPES else None
+    env_norm = os.environ.get("GPU_TYPE", "").strip().lower()
+    if env_norm:
+        return env_norm if env_norm in _AMD_GPU_TYPES else None
+    detected = (_autodetect_gpu_type() or "").strip().lower()
+    return detected if detected in _AMD_GPU_TYPES else None
